+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH6D0A2PG1BDB3D9J68EBGGP"
+        "01KH6DHF8RVABJEGARY62AFYHS"
     )
     r"""The unique identifier of the data source"""