+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-11T13:07:01.572Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-11T13:16:24.133Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")