+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KH6D09XH5ZHKGR6EZ5ZEM5Q2"
+    id: Optional[str] = "01KH6DHF4D9S0Y4R2HHB0TMH45"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KH6D09XBAQRAS93ZMTZTRC47"
+        "tool_01KH6DHF3ZP9ZSJM4KNG4MW8S6"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KH6D09XAXV9FSV28PKWR60XS"
+    id: Optional[str] = "01KH6DHF3ZBZ947GAA8RADBCM8"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KH6D09X9QVY53WAD5REY75P8"
+        "tool_01KH6DHF3XE9P0YXFP0Z0BV4ZR"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KH6D09X0XPNZJK4GXFKDB1A4"
+        "tool_01KH6DHF3W15EXDH2MW3N02N5Z"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KH6D09WZVYPYK9NVWE1JSQQF"
+        "tool_01KH6DHF3TTVTVPSXZ0Q2XYB2X"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KH6D09WY1AERZP2MC3J0AX5K"
+        "tool_01KH6DHF3STMHV3BPA9TCDEWM2"
     )
 
     display_name: Optional[str] = None