+++ packages/orq-rc/RELEASES.md
@@ -5529,3 +5529,13 @@ Based on:
 - [python v4.3.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.53 - packages/orq-rc
+
+## 2026-02-11 13:18:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.710.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.54] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.54 - packages/orq-rc